requirements.txt
openenv-core>=0.2.1
fastapi
uvicorn
pydantic
requests